chore: move LayoutListPage to parent directory
